Add check shim + support
The check shim was lifted mostly as-is from the pg021-bluespec repo,
with minor changes (mkDReg -> mkDWire and printing changes).
HWCRYPTO_CHERI enables cheri support
HWCRYPTO_CHERI_INT_CHECK enables internal checking (ie not using the
shim)
HWCRYPTO_CHERI_FAT decompresses capabilities when they are written and
stores the decompressed format in the registers. Otherwise capabilities
are stored in the compressed (in-memory) format.